server/requirements.txt
flask
opencv-python
numpy
Pillow
flask-cors